_curse/diploma/requirements.txt
aiokafka~=0.11.0
alembic~=1.13.1
asyncpg~=0.29.0
authlib~=1.3.0
bcrypt~=3.1.7
click~=8.1.7
g4f~=0.3.3.0
google-cloud-aiplatform~=1.70.0
google-generativeai==0.8.3
fastapi~=0.111.0
fastapi-pagination~=0.12.30
httpx~=0.27.0
httptools~=0.6.1
itsdangerous
Jinja2~=3.1.4
passlib~=1.7.4
pip~=24.2
protobuf~=5.28.2
psycopg2-binary
pydantic~=2.7.1
pydantic_core~=2.18.4
python-dotenv~=1.0.1
python-jose~=3.3
PyYAML~=6.0.2
redis==5.1.1
requests~=2.32.2
SQLAlchemy~=2.0.30
starlette~=0.37.2
uvicorn==0.30.5
uvloop==0.20.0